# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "golang-company-api",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/api/v1/company/continent/{continent}": {
            "get": {
                "description": "This endpoint allows you to get a company by continent",
                "tags": [
                    "Company"
                ],
                "summary": "Get company",
                "parameters": [
                    {
                        "type": "string",
                        "description": "continent",
                        "name": "command",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_Lucasma95_golang-company-api_src_api_core_entities.CompanyDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/v1/company/country/{country}": {
            "get": {
                "description": "This endpoint allows you to get a company by id",
                "tags": [
                    "Company"
                ],
                "summary": "Get company",
                "parameters": [
                    {
                        "type": "string",
                        "description": "company_id",
                        "name": "command",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_Lucasma95_golang-company-api_src_api_core_entities.CompanyDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/company": {
            "post": {
                "description": "This endpoint allows you to create a company",
                "tags": [
                    "Company"
                ],
                "summary": "Create Company",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Lucasma95_golang-company-api_src_api_http_contracts_company.CreateCompanyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/country": {
            "post": {
                "description": "This endpoint allows you to create a country",
                "tags": [
                    "Country"
                ],
                "summary": "Create Country",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Lucasma95_golang-company-api_src_api_http_contracts_country.CreateCountryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/health": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "This endpoint is used to verify if the api is running",
                "operationId": "health",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/subscription/config/{company_id}": {
            "get": {
                "description": "This endpoint allows you to get a company by id",
                "tags": [
                    "Company"
                ],
                "summary": "Get company",
                "parameters": [
                    {
                        "type": "string",
                        "description": "company_id",
                        "name": "command",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Lucasma95_golang-company-api_src_api_core_entities.Company"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_Lucasma95_golang-company-api_src_api_core_entities.Company": {
            "type": "object",
            "properties": {
                "country": {
                    "$ref": "#/definitions/github_com_Lucasma95_golang-company-api_src_api_core_entities.Country"
                },
                "countryName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_Lucasma95_golang-company-api_src_api_core_entities.CompanyDTO": {
            "type": "object",
            "properties": {
                "continent": {
                    "type": "string"
                },
                "country_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_Lucasma95_golang-company-api_src_api_core_entities.Country": {
            "type": "object",
            "properties": {
                "continent": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_Lucasma95_golang-company-api_src_api_http_contracts_company.CreateCompanyRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "country_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_Lucasma95_golang-company-api_src_api_http_contracts_country.CreateCountryRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "continent": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}